+++ lib/safe-redis-sliding-window-rate-limiter.d.ts
@@ -5,13 +5,13 @@ import { RedisSlidingWindowRateLimiter, RedisSlidingWindowRateLimiterOptions } f
 
 export interface SafeRedisSlidingWindowRateLimiterOptions extends RedisSlidingWindowRateLimiterOptions {
   safe?: true
-  reconnectTimeout?: number
+  reuseRedisAfter?: number
   defaultResponse?: number
 }
 
 export declare class SafeRedisSlidingWindowRateLimiter extends RedisSlidingWindowRateLimiter implements SlidingWindowRateLimiterBackend {
   readonly options: SafeRedisSlidingWindowRateLimiterOptions
-  readonly reconnectTimeout: number
+  readonly reuseRedisAfter: number
   readonly defaultResponse: number
 
   constructor (options?: SafeRedisSlidingWindowRateLimiterOptions)